+++ chaekit-spring/src/main/java/qwerty/chaekit/domain/ebook/EbookRepository.java
@@ -1,4 +1,4 @@
-package qwerty.chaekit.domain.Member.ebook;
+package qwerty.chaekit.domain.ebook;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
