+++ chaekit-spring/src/main/java/qwerty/chaekit/domain/member/user/UserProfileRepository.java
@@ -1,4 +1,4 @@
-package qwerty.chaekit.domain.Member.user;
+package qwerty.chaekit.domain.member.user;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
