+++ chaekit-spring/src/main/java/qwerty/chaekit/global/init/AdminInitializer.java
@@ -6,11 +6,10 @@ import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
-import qwerty.chaekit.domain.Member.Member;
-import qwerty.chaekit.domain.Member.MemberRepository;
-import qwerty.chaekit.domain.Member.enums.Role;
-import qwerty.chaekit.domain.Member.publisher.PublisherProfile;
-import qwerty.chaekit.domain.Member.publisher.PublisherProfileRepository;
+import qwerty.chaekit.domain.member.Member;
+import qwerty.chaekit.domain.member.enums.Role;
+import qwerty.chaekit.domain.member.publisher.PublisherProfile;
+import qwerty.chaekit.domain.member.publisher.PublisherProfileRepository;
 import qwerty.chaekit.global.exception.BadRequestException;
 import qwerty.chaekit.global.properties.AdminProperties;
 import qwerty.chaekit.service.AdminService;
