+++ chaekit-spring/src/main/java/qwerty/chaekit/service/EbookFileService.java
@@ -4,9 +4,9 @@ import jakarta.persistence.EntityManager;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-import qwerty.chaekit.domain.Member.ebook.Ebook;
-import qwerty.chaekit.domain.Member.ebook.EbookRepository;
-import qwerty.chaekit.domain.Member.publisher.PublisherProfile;
+import qwerty.chaekit.domain.ebook.Ebook;
+import qwerty.chaekit.domain.ebook.EbookRepository;
+import qwerty.chaekit.domain.member.publisher.PublisherProfile;
 import qwerty.chaekit.dto.upload.EbookDownloadResponse;
 import qwerty.chaekit.dto.upload.EbookUploadRequest;
 import qwerty.chaekit.global.exception.BadRequestException;
