+++ chaekit-spring/src/main/java/qwerty/chaekit/service/PublisherJoinService.java
@@ -3,10 +3,10 @@ package qwerty.chaekit.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import qwerty.chaekit.domain.Member.Member;
-import qwerty.chaekit.domain.Member.enums.Role;
-import qwerty.chaekit.domain.Member.publisher.PublisherProfile;
-import qwerty.chaekit.domain.Member.publisher.PublisherProfileRepository;
+import qwerty.chaekit.domain.member.Member;
+import qwerty.chaekit.domain.member.enums.Role;
+import qwerty.chaekit.domain.member.publisher.PublisherProfile;
+import qwerty.chaekit.domain.member.publisher.PublisherProfileRepository;
 import qwerty.chaekit.dto.PublisherJoinRequest;
 import qwerty.chaekit.dto.PublisherJoinResponse;
 import qwerty.chaekit.global.exception.BadRequestException;
