+++ chaekit-spring/src/main/java/qwerty/chaekit/service/PublisherService.java
@@ -3,8 +3,8 @@ package qwerty.chaekit.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import qwerty.chaekit.domain.Member.publisher.PublisherProfile;
-import qwerty.chaekit.domain.Member.publisher.PublisherProfileRepository;
+import qwerty.chaekit.domain.member.publisher.PublisherProfile;
+import qwerty.chaekit.domain.member.publisher.PublisherProfileRepository;
 import qwerty.chaekit.dto.PublisherMemberResponse;
 import qwerty.chaekit.global.exception.NotFoundException;
 import qwerty.chaekit.global.security.resolver.LoginMember;
