+++ chaekit-spring/src/main/java/qwerty/chaekit/service/UserJoinService.java
@@ -3,10 +3,10 @@ package qwerty.chaekit.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import qwerty.chaekit.domain.Member.Member;
-import qwerty.chaekit.domain.Member.enums.Role;
-import qwerty.chaekit.domain.Member.user.UserProfile;
-import qwerty.chaekit.domain.Member.user.UserProfileRepository;
+import qwerty.chaekit.domain.member.Member;
+import qwerty.chaekit.domain.member.enums.Role;
+import qwerty.chaekit.domain.member.user.UserProfile;
+import qwerty.chaekit.domain.member.user.UserProfileRepository;
 import qwerty.chaekit.dto.UserJoinRequest;
 import qwerty.chaekit.dto.UserJoinResponse;
 import qwerty.chaekit.global.exception.BadRequestException;
