+++ chaekit-spring/src/main/java/qwerty/chaekit/service/UserService.java
@@ -3,7 +3,7 @@ package qwerty.chaekit.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import qwerty.chaekit.domain.Member.user.UserProfileRepository;
+import qwerty.chaekit.domain.member.user.UserProfileRepository;
 import qwerty.chaekit.dto.UserMemberResponse;
 import qwerty.chaekit.global.exception.NotFoundException;
 import qwerty.chaekit.global.security.resolver.LoginMember;
